% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{analyze_removed_questions}
\alias{analyze_removed_questions}
\title{Analyze effect of removal of questions on accuracy with Random Forest.}
\usage{
analyze_removed_questions(data, imp_num, party_col)
}
\arguments{
\item{data}{The dataset with questions and party affiliation.}

\item{imp_num}{Question numbers from least important to most important as vector.}

\item{party_col}{Column name of party affiliation.}
}
\value{
Data frame with class error by removed question.
}
\description{
Removes questions based on their importance value (starting from least important),
runs Random Forest, and analyzes accuracy.
}
