% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{classComp}
\alias{classComp}
\title{Predict party membership with caret}
\usage{
res <- classComp(data, 10, 1, "rf", "party")
}
\arguments{
\item{data}{The data set to be used: should only include party column and predictors!}

\item{k}{Selected k for cross-validation.}

\item{model}{The caret model to be used in prediction.}

\item{party_col}{Column name that includes party affiliation.}

\item{repeats}{Number of repeats for CV (increases stability).}
}
\value{
Data frame that includes optimized parameters, accuracy and kappa information.
}
\description{
Perform k-fold CV and hyperparameter optimization with caret to predict party affiliation based on 
election machine questions.
}
