% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.get_HS_2015_data}
\alias{.get_HS_2015_data}
\title{Get the HS data}
\usage{
.get_HS_2015_data()
}
\value{
Tbl dataframe.
}
\description{
Loads the HS data set and does a little preprocessing. The dataset is not supplied with the package,
because it is not public.
}
