% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elecmachine-package.R
\docType{package}
\name{elecmachine-package}
\alias{elecmachine}
\alias{elecmachine-package}
\title{elecmachine: Scripts for analysing Finnish Election Machine datasets}
\description{
Functions that can be used to analyze the election machine data from Helsingin Sanomat (year 2015) and YLE (years 2011 and 2015).
}
\author{
\strong{Maintainer}: Tommi Pajala \email{tjpajala@gmail.com}

}
\keyword{internal}
