% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{error_ggplot}
\alias{error_ggplot}
\title{Plot classwise error by removed question with ggplot.}
\usage{
error_ggplot(res)
}
\arguments{
\item{res}{Data frame returned by \code{\link{analyze_removed_questions}}}
}
\description{
Plot classwise error by removed question with ggplot.
}
