% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{FA_ggplot}
\alias{FA_ggplot}
\title{Plot factor plot with ggplot}
\usage{
FA_ggplot(PAF(data, 2, FALSE, names(select(data, q1:q30))),flip=20)
}
\arguments{
\item{fa}{The data frame including factor scores.}

\item{flip}{Degrees to rotate axes (so that leftists are on the left, etc.) Default value 20 (works for HS data).}
}
\description{
Plots points on a 2D coordinate scatter plot, based on their factor scores.
}
