% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{FAplot}
\alias{FAplot}
\title{Plot party candidates on a 2D factor plot}
\usage{
FAplot(PAF(data, 2, FALSE, names(select(data, q1:q30))), centers=FALSE, add=FALSE)
}
\arguments{
\item{fa_ans}{The data set and factor scores together. Use \code{\link{PAF}} to get this.}

\item{party_col}{Column name that represents party (as string).}

\item{centers}{Boolen to control for drawing party centers.}

\item{add}{Do you want to add points to a previous plot, or draw a new one?}

\item{pch}{Which plot symbol to use.}

\item{flip}{Degrees of rotation for the plot, can be used to align left-right axis in the usual political direction}

\item{...}{Any additional arguments for the plot function}
}
\description{
Takes candidate answers and their factor scores, and plots them on a 2D plot (that can be rotated), 
coloring candidates by party.
}
