% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_data_cols}
\alias{get_data_cols}
\title{Get the question columns for the specified data set}
\usage{
q_cols <- get_data_cols('yle_2011', data)
}
\arguments{
\item{dataset_name}{Name of the dataset ('yle_2011', 'yle_2015' or 'hs_2015')}
}
\value{
List of columns that represent the questions.
}
\description{
Get the question columns for the specified data set
}
