% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_dataset}
\alias{get_dataset}
\title{Get the required data set}
\usage{
get_dataset(name, filter_precinct = NULL)
}
\arguments{
\item{name}{Dataset name (options: "yle_2011", "yle_2015", "hs_2015", "yle_2019")}
}
\value{
Tbl dataframe.
}
\description{
Convenience function for loading the required dataset.
}
