% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{get_text_for_variance_nmost}
\alias{get_text_for_variance_nmost}
\title{Get text for the questions with most or least variance}
\usage{
get_text_for_variance_nmost(data, q_cols, 5, -1, "var")
}
\arguments{
\item{data}{Dataset}

\item{q_cols}{Question columns}

\item{nmost}{(Optional.) How many questions to return? Default: 5}

\item{direction}{(Optional.) Either most(\code{direction=1}) or least (\code{direction=-1}) variance. Default: -1}

\item{function_to_use}{(Optional.) What function to use for analysis of variance? Default: "var"}
}
\description{
Provides the question text for the \code{nmost} questions with either most(\code{direction=1}) or
least (\code{direction=-1}) variance.
}
