% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{PAF}
\alias{PAF}
\title{Principal Axis Factoring for election machine answers}
\usage{
paf <- PAF(data, 3, FALSE, cols=names(select(data,q1:q30)))
}
\arguments{
\item{data}{The data set to be used.}

\item{nfactors}{How many factors to use in the analysis.}

\item{vss}{Should a VSS analysis be run to analyze the amount of factors?}

\item{cols}{List of column names that include the data to be analysed.}
}
\value{
A list that includes the factor original data augmented with scores, correlations, determinant of the 
correlation matrix, KMO score, and factor loadings. See \code{\link[psych]{fa} for more
information.}
}
\description{
Takes the answers in the election machine, and completes
Principal Axis Factoring (PAF) for them. Returns factor
scores, correlations, the determinant, KMO score and 
factor loadings.
}
