% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_for_all_questions}
\alias{plot_for_all_questions}
\title{Violin plot for distributions of candidates across questions}
\usage{
plot_for_all_questions(data, q_cols)
}
\arguments{
\item{data}{Data set to be used}

\item{q_cols}{The columns defining questions}
}
\description{
Violin plots that show how candidates are distributed across questions (optionally across parties).
}
