% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_single_question}
\alias{plot_single_question}
\title{Plot for a single question and all parties}
\usage{
plot_single_question(data, 5, q_cols, jitter=TRUE)
}
\arguments{
\item{data}{Dataset.}

\item{q_num}{Question to be analyzed.}

\item{q_cols}{The columns defining questions}
}
\description{
Plot showing distribution of candidates of different parties for a single question
}
\details{
Plots the distribution of candidates of different parties for a single question, as a combination
of tile and jitter plots.
}
