% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare data for analysis}
\usage{
data <- prepare_data(data, q_cols, party_col)
}
\arguments{
\item{data}{Dataset.}

\item{q_cols}{Question columns. See \code{\link{get_data_cols}}.}
}
\value{
Dataset with subbed and factored party columns, and numeric question cols.
}
\description{
Prepares data for analysis by combining small parties, factoring the party column, 
and ensuring that question columns are numeric.
}
