% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{removeQname}
\alias{removeQname}
\title{Remove election machine question with name \code{qname}.}
\usage{
removeQname(data, qname)
}
\arguments{
\item{data}{Dataset}

\item{qname}{The name of question to be removed}
}
\value{
Dataset without the removed question.
}
\description{
Remove election machine question with name \code{qname}.
}
