% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{set_small_parties_to_other}
\alias{set_small_parties_to_other}
\title{Combine small parties to group "Other"}
\usage{
df$party <- set_small_parties_to_other(df, "party")
}
\arguments{
\item{data}{Data set to be used (either tbl or dataframe)}

\item{colname_party}{Column name that has the party information (default is "party")}

\item{limit}{Candidate limit for setting party membership to "Other". Default: 10}
}
\value{
Returns the party column
}
\description{
Takes all parties with less than \code{nlimit} members, and changes their party
to "Other".
}
