% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sub_parties_for_shortcodes}
\alias{sub_parties_for_shortcodes}
\title{Substitute Finnish Party names with short versions.}
\usage{
data$party <- sub_parties_for_shortcodes(datacol=dataparty)
}
\arguments{
\item{datacol}{Vector of parties to be replaced.}
}
\description{
Takes all party names and substitutes their short versions.
}
