% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{table_question_variance}
\alias{table_question_variance}
\title{Produce table about question variance}
\usage{
table_question_variance(data, q_cols)
}
\arguments{
\item{data}{Dataset.}

\item{q_cols}{Question columns}

\item{cols_to_analyze}{(Optional.) Only a subset of q_cols, if that's all you want to look at. Default is all columns.}

\item{functions_to_use}{(Optional.) What functions to analyze across parties? Default is "var".}
}
\value{
A table in tbl format.
}
\description{
Produces a table showcasing within-party and between-party variance for all the questions.
}
