% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_YLE_2011_data}
\alias{get_YLE_2011_data}
\title{Get the YLE 2011 candidate data}
\usage{
get_YLE_2011_data(filter_precinct = NULL)
}
\arguments{
\item{filter_precinct}{Vector of precincts to include. Defaults to NULL.}
}
\value{
Tbl dataframe.
}
\description{
Fetches and preprocesses the YLE 2011 candidate data.
}
