% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_YLE_2015_data}
\alias{get_YLE_2015_data}
\title{Get the YLE 2015 candidate data}
\usage{
get_YLE_2015_data()
}
\value{
Tbl dataframe.
}
\description{
Fetches and preprocesses the YLE 2015 candidate data.
}
